# src/docs/swagger.yaml
openapi: 3.1.0
info:
  title: Budget API
  version: 1.0.0
  description: Budget API Information
  contact:
    name: Adylsha Yumayev
servers:
  - url: http://localhost:3000
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        ID:
          type: integer
        EMAIL:
          type: string
          format: email
        FIRST_NAME:
          type: string
        LAST_NAME:
          type: string
        CREATED_AT:
          type: string
          format: date-time
        UPDATED_AT:
          type: string
          format: date-time
      required:
        - EMAIL
        - FIRST_NAME
        - LAST_NAME
...
paths:
...
  /users:
    get:
      summary: Get all users
      tags: [Users]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Only admins can access this route
  /users/{id}:
    get:
      summary: Get user by ID
      description: Can only retrieve the current authenticated user's data
      tags: [Users]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/User'
        '403':
          description: Forbidden - Can only access your own data
        '404':
          description: User not found

    put:
      summary: Update user
      description: Can only update the current authenticated user's data
      tags: [Users]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                EMAIL:
                  type: string
                  format: email
                FIRST_NAME:
                  type: string
                LAST_NAME:
                  type: string
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/User'
        '403':
          description: Forbidden - Can only update your own data
        '404':
          description: User not found

    delete:
      summary: Delete user
      description: Can only delete the current authenticated user's account
      tags: [Users]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: null
        '403':
          description: Forbidden - Can only delete your own account
        '404':
          description: User not found

  #
  # CATEGORY ROUTES
  #
...